Renamed Plugins
- Renamed Plugins for clarity and to reduce redundancy
