+++ src/plugins/pregame.rs
